Added a working Classic / Gutenberg switch option in the editing menu.
