+++ WordPress/Classes/Utility/Editor/EditorFactory.swift
@@ -11,11 +11,15 @@ class EditorFactory {
 
     // MARK: - Editor: Instantiation
 
-    func instantiateEditor(for post: AbstractPost) -> UIViewController & PostEditor {
+    func instantiateEditor(
+        for post: AbstractPost,
+        switchToAztec: @escaping (UIViewController & PostEditor) -> (),
+        switchToGutenberg: @escaping (UIViewController & PostEditor) -> ()) -> UIViewController & PostEditor {
+
         if gutenbergSettings.mustUseGutenberg(for: post) {
-            return GutenbergViewController(post: post)
+            return GutenbergViewController(post: post, switchToAztec: switchToAztec)
         } else {
-            return AztecPostViewController(post: post)
+            return AztecPostViewController(post: post, switchToGutenberg: switchToGutenberg)
         }
     }
 }
