+++ WordPress/Classes/ViewRelated/Gutenberg/GutenbergViewController.swift
@@ -40,6 +40,10 @@ class GutenbergViewController: UIViewController, PostEditor {
         return view
     }()
 
+    // MARK: - Aztec
+
+    private let switchToAztec: (UIViewController & PostEditor) -> ()
+
     // MARK: - PostEditor
 
     var html: String {
@@ -135,11 +139,15 @@ class GutenbergViewController: UIViewController, PostEditor {
     private var requestHTMLReason: RequestHTMLReason?
 
     // MARK: - Initializers
-    required init(post: AbstractPost) {
+    required init(
+        post: AbstractPost,
+        switchToAztec: @escaping (UIViewController & PostEditor) -> ()) {
+
         self.post = post
         self.gutenberg = Gutenberg(props: ["initialData": self.post.content ?? ""])
         self.verificationPromptHelper = AztecVerificationPromptHelper(account: self.post.blog.account)
         self.shouldRemovePostOnDismiss = post.hasNeverAttemptedToUpload()
+        self.switchToAztec = switchToAztec
 
         super.init(nibName: nil, bundle: nil)
         self.postTitle = post.postTitle ?? ""
@@ -332,5 +340,113 @@ private extension GutenbergViewController {
         static let titleTextFieldHeight: CGFloat = 50.0
         static let titleTextFieldLeftPadding: CGFloat = 10.0
         static let titleTextFieldBottomSeparatorHeight: CGFloat = 1.0
+    }
+}
+
+// MARK: - MoreActions
+
+/// This extension handles the "more" actions triggered by the top right
+/// navigation bar button of Gutenberg editor.
+extension GutenbergViewController {
+
+    private enum ErrorCode: Int {
+        case expectedSecondaryAction = 1
+    }
+
+    func displayMoreSheet() {
+        let alert = UIAlertController(title: nil, message: nil, preferredStyle: .actionSheet)
+
+        //TODO: Comment in when bridge is ready
+        /*if mode == .richText {
+         // NB : This is a candidate for plurality via .stringsdict, but is limited by https://github.com/wordpress-mobile/WordPress-iOS/issues/6327
+         let textCounterTitle = String(format: NSLocalizedString("%li words, %li characters", comment: "Displays the number of words and characters in text"), richTextView.wordCount, richTextView.characterCount)
+         
+         alert.title = textCounterTitle
+         }*/
+
+        if postEditorStateContext.isSecondaryPublishButtonShown,
+            let buttonTitle = postEditorStateContext.secondaryPublishButtonText {
+
+            alert.addDefaultActionWithTitle(buttonTitle) { _ in
+                self.secondaryPublishButtonTapped()
+            }
+        }
+
+        //TODO: Comment in when bridge is ready
+        /*let toggleModeTitle: String = {
+         if mode == .richText {
+         return MoreSheetAlert.htmlTitle
+         } else {
+         return MoreSheetAlert.richTitle
+         }
+         }()
+         
+         alert.addDefaultActionWithTitle(toggleModeTitle) { [unowned self] _ in
+         self.toggleEditingMode()
+         }*/
+
+        alert.addDefaultActionWithTitle(MoreSheetAlert.classicTitle) { [unowned self] _ in
+            self.switchToAztec(self)
+        }
+
+        alert.addDefaultActionWithTitle(MoreSheetAlert.previewTitle) { [weak self] _ in
+            self?.displayPreview()
+        }
+
+        //TODO: Comment in when bridge is ready
+        /*
+         if Feature.enabled(.revisions) && (post.revisions ?? []).count > 0 {
+         alert.addDefaultActionWithTitle(MoreSheetAlert.historyTitle) { [weak self] _ in
+         self?.displayHistory()
+         }
+         }*/
+
+        alert.addDefaultActionWithTitle(MoreSheetAlert.postSettingsTitle) { [weak self] _ in
+            self?.displayPostSettings()
+        }
+
+        alert.addCancelActionWithTitle(MoreSheetAlert.keepEditingTitle)
+
+        alert.popoverPresentationController?.barButtonItem = navigationBarManager.moreBarButtonItem
+
+        present(alert, animated: true)
+    }
+
+    func secondaryPublishButtonTapped() {
+        guard let action = self.postEditorStateContext.secondaryPublishButtonAction else {
+            // If the user tapped on the secondary publish action button, it means we should have a secondary publish action.
+            let error = NSError(domain: errorDomain, code: ErrorCode.expectedSecondaryAction.rawValue, userInfo: nil)
+            Crashlytics.sharedInstance().recordError(error)
+            return
+        }
+
+        let secondaryStat = self.postEditorStateContext.secondaryPublishActionAnalyticsStat
+
+        let publishPostClosure = { [unowned self] in
+            self.publishPost(
+                action: action,
+                dismissWhenDone: action.dismissesEditor,
+                analyticsStat: secondaryStat)
+        }
+
+        if presentedViewController != nil {
+            dismiss(animated: true, completion: publishPostClosure)
+        } else {
+            publishPostClosure()
+        }
+    }
+}
+
+// MARK: - Constants
+
+extension GutenbergViewController {
+    private struct MoreSheetAlert {
+        static let classicTitle = NSLocalizedString("Switch to Classic Editor", comment: "Switches from Gutenberg mobile to the Classic editor")
+        static let htmlTitle = NSLocalizedString("Switch to HTML Mode", comment: "Switches the Editor to HTML Mode")
+        static let richTitle = NSLocalizedString("Switch to Visual Mode", comment: "Switches the Editor to Rich Text Mode")
+        static let previewTitle = NSLocalizedString("Preview", comment: "Displays the Post Preview Interface")
+        static let historyTitle = NSLocalizedString("History", comment: "Displays the History screen from the editor's alert sheet")
+        static let postSettingsTitle = NSLocalizedString("Post Settings", comment: "Name of the button to open the post settings")
+        static let keepEditingTitle = NSLocalizedString("Keep Editing", comment: "Goes back to editing the post.")
     }
 }
