+++ WordPress/Classes/ViewRelated/Post/PostEditor.swift
@@ -4,9 +4,6 @@ import UIKit
 /// Common interface to all editors
 ///
 protocol PostEditor: class, UIViewControllerTransitioningDelegate {
-    /// Initialize editor with a post.
-    ///
-    init(post: AbstractPost)
 
     /// The post being edited.
     ///
